requirements.txt
flask==2.0.1
pandas==1.3.3
spacy==3.2.0
faiss-cpu==1.7.2
numpy==1.21.2
sentence-transformers==2.2.2